# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Job Scout API
  version: 1.0.0
  description: |
    Internal API for the job-scout job aggregator. This spec is the contract:
    the Spring Boot backend implements API interfaces generated from this file,
    and the React frontend's types are regenerated from it.
  contact:
    name: job-scout
    url: https://github.com/halitdincer/job-scout
  license:
    name: MIT
    identifier: MIT
servers:
  - url: /api/v1
    description: Default base URL (versioned)
security:
  - cookieAuth: []
...
paths:
...
  /auth/me:
    get:
      tags: [Auth]
      operationId: getCurrentUser
      summary: Get the currently-authenticated user
      description: Returns basic identity information for the user associated with the current session.
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CurrentUser' }
        '401': { $ref: '#/components/responses/Unauthorized' }

  # ------------------------------------------------------------------
  # Sources
  # ------------------------------------------------------------------
...
  /jobs:
    get:
      tags: [Jobs]
      operationId: listJobs
      summary: List job listings with filter, sort, and pagination
      description: |
        The core jobs endpoint. Supports a recursive filter DSL, multi-field
        sorting (including the user-specific `seen` sort), and page-based
        pagination.
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - $ref: '#/components/parameters/SortParam'
        - $ref: '#/components/parameters/FilterParam'
      responses:
        '200':
          description: Paginated job listings
          content:
            application/json:
              schema: { $ref: '#/components/schemas/JobListingPage' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
  /jobs/facets:
    get:
      tags: [Jobs]
      operationId: getJobFacets
      summary: Faceted aggregations for jobs
      description: |
        Returns distinct value counts for the requested facet fields. Filters
        from the same query are applied (so facets reflect the current view).
      parameters:
        - $ref: '#/components/parameters/FilterParam'
        - name: fields
          in: query
          required: true
          description: Comma-separated list of facet fields to aggregate.
          schema:
            type: array
            items:
              type: string
              enum: [source_name, status, country, region, city]
          style: form
          explode: false
      responses:
        '200':
          description: Facet buckets
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FacetsResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /jobs/{id}/seen:
    post:
      tags: [Jobs]
      operationId: markJobSeen
      summary: Mark a job as seen by the current user
      description: Idempotent — re-marking has no effect.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200':
          description: Mark recorded (or already present)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MarkSeenResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ------------------------------------------------------------------
  # Locations
  # ------------------------------------------------------------------
...
  /runs:
    get:
      tags: [Runs]
      operationId: listRuns
      summary: List ingestion run history (most recent first)
      description: Returns a paginated list of ingestion runs for the current user, ordered by newest run first.
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
      responses:
        '200':
          description: Paginated runs
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RunPage' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [Runs]
      operationId: triggerRun
      summary: Trigger a new ingestion run
      description: Spawns ingestion asynchronously. Returns the created `Run` immediately with status PENDING.
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TriggerRunRequest' }
      responses:
        '202':
          description: Run accepted
          headers:
            Location:
              description: URL of the created run resource
              schema: { type: string, format: uri }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Run' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /runs/{id}:
    get:
      tags: [Runs]
      operationId: getRun
      summary: Get a single run
      description: Returns one ingestion run by id when it belongs to the current user.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200':
          description: Run
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Run' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ------------------------------------------------------------------
  # SavedViews
  # ------------------------------------------------------------------
  /views:
    get:
      tags: [SavedViews]
      operationId: listSavedViews
      summary: List the current user's saved views
      description: Returns saved views owned by the current user, ordered by name.
      responses:
        '200':
          description: Saved views (ordered by name)
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/SavedView' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [SavedViews]
      operationId: createSavedView
      summary: Create a new saved view for the current user
      description: Persists the current user's filter, sort, column, and page-size preferences as a named saved view.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SavedViewCreateRequest' }
      responses:
        '201':
          description: Saved view created
          headers:
            Location:
              description: URL of the created saved view
              schema: { type: string, format: uri }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SavedView' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
  /views/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: integer, format: int64 }
    get:
      tags: [SavedViews]
      operationId: getSavedView
      summary: Get a saved view by id
      description: Returns one saved view by id when it belongs to the current user.
      responses:
        '200':
          description: Saved view
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SavedView' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [SavedViews]
      operationId: updateSavedView
      summary: Replace a saved view
      description: Replaces a saved view owned by the current user with the supplied preferences.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SavedViewUpdateRequest' }
      responses:
        '200':
          description: Updated saved view
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SavedView' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
    delete:
      tags: [SavedViews]
      operationId: deleteSavedView
      summary: Delete a saved view
      description: Deletes a saved view owned by the current user.
      responses:
        '204':
          description: Deleted
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
components:
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: JSESSIONID
      description: Spring Security session cookie

  # ==================================================================
  # Reusable parameters
  # ==================================================================
  parameters:
    PageParam:
      name: page
      in: query
      description: Zero-based page index.
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
    PageSizeParam:
      name: pageSize
      in: query
      description: Page size.
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 250
        default: 50
    SortParam:
      name: sort
      in: query
      description: |
        Multi-field sort. Comma-separated `field:dir` tokens, e.g.
        `first_seen_at:desc,title:asc`. The pseudo-field `seen` sorts by
        whether the current user has marked the listing as seen.
      required: false
      style: form
      explode: false
      schema:
        type: array
        items: { type: string, pattern: '^[a-z_]+(:asc|:desc)?$' }
    FilterParam:
      name: filter
      in: query
      description: |
        Recursive filter DSL, JSON-encoded. See `FilterExpression`.
        Example: `{"and":[{"field":"status","op":"eq","value":"ACTIVE"},{"field":"published_at","op":"in_last_days","value":7}]}`
      required: false
      schema:
        type: string

  # ==================================================================
  # Reusable error responses
  # ==================================================================
  responses:
    BadRequest:
      description: Malformed request
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Unauthorized:
      description: Authentication required
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    NotFound:
      description: Resource not found
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    UnprocessableEntity:
      description: Validation failed
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }

  # ==================================================================
  # Schemas
  # ==================================================================
  schemas:
...
    CurrentUser:
      type: object
      required: [id, username, isStaff, isSuperuser]
      properties:
        id:          { type: integer, format: int64 }
        username:    { type: string }
        firstName:   { type: string }
        lastName:    { type: string }
        email:       { type: string, format: email }
        isStaff:     { type: boolean }
        isSuperuser: { type: boolean }
    Problem:
      description: RFC 7807 Problem Details.
      type: object
      required: [type, title, status]
      properties:
        type:    { type: string, format: uri, default: 'about:blank' }
        title:   { type: string }
        status:  { type: integer, minimum: 400, maximum: 599 }
        detail:  { type: string }
        instance: { type: string, format: uri }
        errors:
          description: Per-field validation errors (for 422 responses).
          type: array
          items:
            type: object
            required: [field, message]
            properties:
              field:   { type: string }
              message: { type: string }
      example:
        type: 'about:blank'
        title: 'Bad Request'
        status: 400
        detail: 'Sort field unknown: foo'

    # ---------- Sources -----------------------------------------------
...
    LocationTag:
      type: object
      required: [id, name]
      properties:
        id:          { type: integer, format: int64 }
        name:        { type: string }
        countryCode: { type: ['string', 'null'], maxLength: 2, minLength: 2 }
        regionCode:  { type: ['string', 'null'], maxLength: 10 }
        city:        { type: ['string', 'null'] }
        geoKey:      { type: ['string', 'null'], description: 'Derived hierarchical key (country / region / city).' }

    # ---------- JobListing ---------------------------------------------
    ListingStatus:
      type: string
      enum: [ACTIVE, EXPIRED]
    JobListing:
      type: object
      required:
        - id
        - sourceId
        - sourceName
        - externalId
        - title
        - url
        - status
        - locations
        - firstSeenAt
        - lastSeenAt
        - seen
      properties:
        id:              { type: integer, format: int64 }
        sourceId:        { type: integer, format: int64 }
        sourceName:      { type: string }
        externalId:      { type: string }
        title:           { type: string }
        url:             { type: string, format: uri }
        status:          { $ref: '#/components/schemas/ListingStatus' }
        locations:
          type: array
          items: { $ref: '#/components/schemas/LocationTag' }
        publishedAt:     { type: ['string', 'null'], format: date-time }
        updatedAtSource: { type: ['string', 'null'], format: date-time }
        firstSeenAt:     { type: string, format: date-time }
        lastSeenAt:      { type: string, format: date-time }
        expiredAt:       { type: ['string', 'null'], format: date-time }
        seen:
          type: boolean
          description: Whether the current user has marked this listing as seen.
    JobListingPage:
      type: object
      required: [items, page, pageSize, total, hasNext]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/JobListing' }
        page:     { type: integer, minimum: 0 }
        pageSize: { type: integer, minimum: 1 }
        total:    { type: integer, minimum: 0 }
        hasNext:  { type: boolean }
    MarkSeenResponse:
      type: object
      required: [id, seen]
      properties:
        id:   { type: integer, format: int64 }
        seen: { type: boolean }

    # ---------- Filter DSL ---------------------------------------------
...
    FilterOperator:
      type: string
      enum:
        - contains
        - not_contains
        - eq
        - neq
        - in
        - not_in
        - is_empty
        - is_not_empty
        - before
        - after
        - in_last_days
    FilterPredicate:
      type: object
      required: [field, op]
      properties:
        field: { $ref: '#/components/schemas/FilterField' }
        op:    { $ref: '#/components/schemas/FilterOperator' }
        value:
          description: Operator value. Null for `is_empty`/`is_not_empty`; array for `in`/`not_in`; integer days for `in_last_days`; string for others.
    FilterExpression:
      description: |
        Recursive filter tree. A node is one of:
        - a single predicate (`{field, op, value}`)
        - `{and: [Expr, ...]}`
        - `{or: [Expr, ...]}`
        - `{not: Expr}`
      oneOf:
        - $ref: '#/components/schemas/FilterPredicate'
        - type: object
          required: [and]
          properties:
            and:
              type: array
              minItems: 1
              items: { $ref: '#/components/schemas/FilterExpression' }
        - type: object
          required: [or]
          properties:
            or:
              type: array
              minItems: 1
              items: { $ref: '#/components/schemas/FilterExpression' }
        - type: object
          required: [not]
          properties:
            not: { $ref: '#/components/schemas/FilterExpression' }

    # ---------- Facets -------------------------------------------------
    FacetBucket:
      type: object
      required: [value, count]
      properties:
        value: { type: string }
        count: { type: integer, minimum: 0 }
    FacetsResponse:
      type: object
      additionalProperties:
        type: array
        items: { $ref: '#/components/schemas/FacetBucket' }
      description: |
        Map of facet field → buckets. Keys are limited to the values
        documented on `getJobFacets`.
      example:
        source_name:
          - { value: 'Acme', count: 42 }
          - { value: 'Beta', count: 7 }
        status:
          - { value: 'ACTIVE', count: 100 }

    # ---------- Runs ---------------------------------------------------
    RunStatus:
      type: string
      enum: [PENDING, RUNNING, COMPLETED, FAILED]
    Run:
      type: object
      required:
        - id
        - status
        - sourcesProcessed
        - listingsCreated
        - listingsUpdated
        - listingsExpired
        - createdAt
      properties:
        id:                { type: integer, format: int64 }
        status:            { $ref: '#/components/schemas/RunStatus' }
        startedAt:         { type: ['string', 'null'], format: date-time }
        finishedAt:        { type: ['string', 'null'], format: date-time }
        sourcesProcessed:  { type: integer, minimum: 0 }
        listingsCreated:   { type: integer, minimum: 0 }
        listingsUpdated:   { type: integer, minimum: 0 }
        listingsExpired:   { type: integer, minimum: 0 }
        errorMessage:      { type: ['string', 'null'] }
        createdAt:         { type: string, format: date-time }
    RunPage:
      type: object
      required: [items, page, pageSize, total, hasNext]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Run' }
        page:     { type: integer, minimum: 0 }
        pageSize: { type: integer, minimum: 1 }
        total:    { type: integer, minimum: 0 }
        hasNext:  { type: boolean }
    TriggerRunRequest:
      type: object
      description: Optional payload to scope a triggered run to a single source.
      properties:
        sourceId: { type: ['integer', 'null'], format: int64 }

    # ---------- SavedViews ---------------------------------------------
    SortSpec:
      type: object
      required: [field, dir]
      properties:
        field: { type: string, pattern: '^[a-z_]+$' }
        dir:   { type: string, enum: [asc, desc] }
    ColumnDef:
      type: object
      required: [id, visible]
      properties:
        id:      { type: string, pattern: '^[a-z_]+$' }
        visible: { type: boolean }
        width:   { type: ['integer', 'null'], minimum: 40, maximum: 800 }
    SavedView:
      type: object
      required: [id, name, columns, sort, createdAt, updatedAt]
      properties:
        id:               { type: integer, format: int64 }
        name:             { type: string, minLength: 1, maxLength: 255 }
        filterExpression: { $ref: '#/components/schemas/FilterExpression' }
        columns:
          type: array
          items: { $ref: '#/components/schemas/ColumnDef' }
        sort:
          type: array
          items: { $ref: '#/components/schemas/SortSpec' }
        config:
          type: object
          additionalProperties: true
          description: 'Free-form per-view config (page size, density, etc.).'
        createdAt:        { type: string, format: date-time }
        updatedAt:        { type: string, format: date-time }
    SavedViewCreateRequest:
      type: object
      required: [name, columns, sort]
      properties:
        name:             { type: string, minLength: 1, maxLength: 255 }
        filterExpression: { $ref: '#/components/schemas/FilterExpression' }
        columns:
          type: array
          minItems: 1
          items: { $ref: '#/components/schemas/ColumnDef' }
        sort:
          type: array
          items: { $ref: '#/components/schemas/SortSpec' }
        config:
          type: object
          additionalProperties: true
    SavedViewUpdateRequest:
      $ref: '#/components/schemas/SavedViewCreateRequest'